requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
elasticsearch==8.13.0
pydantic
python-dotenv
